+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Fixed
+- Avoid parsing stderr within LocalExecutor class.
 
 ## [0.4.2][changes-0.4.2] - 2025-12-15
 ### Fixed